Stats drawn in appropriate places for spectator
